Use tox to declare the test matrix
This eases the burden of locally testing the plugin across
all supported python and pytest versions, and will simplify
some future exclusions in the matrix (e.g. pytest<6.2
incompatibility with python-3.10)

CI jobs are simplified to be agnostic of the underlying test matrix,
restricting each job to the python versions they explicitly
setup with a negative lookup in the TOX_SKIP_ENV variable.
Each job will be larger due to less venv fragmentation but they
will be collectively more efficient thanks to the reutilization of
the pulled and built docker images.

Also introduce support for pytest-7 and move pytest configuration to the
new tox.ini file, as this is preferable instead of the current setup.cfg
https://docs.pytest.org/en/6.2.x/reference.html#configuration-options